hacked synapse model so that each variable has synapse name in it. hacked VSneuron model to add one current variable for each incoming synapse, both DL-Int-1 and JODLInt1DlInt2 working

+++ brianUtils.py
@@ -3,4 +3,4 @@ from brian2.units.fundamentalunits import Quantity
 
 def getSimT(simDur: Quantity, simStepSize: Quantity) -> Quantity:
 
-    return np.arange(int(simDur / simStepSize)) * simStepSize
+    return np.arange(simDur / simStepSize) * simStepSize
